+++ src/components/tabs/__internal__/tab-title/index.ts
@@ -1,2 +1,2 @@
-export { default, TabTitleContext } from "./tab-title.component";
+export { default } from "./tab-title.component";
 export type { TabTitleProps } from "./tab-title.component";
