+++ src/components/tabs/__internal__/tab-title/tab-title.context.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+interface TabTitleContextProps {
+  isInTab?: boolean;
+}
+
+export default React.createContext<TabTitleContextProps>({});
