+++ src/components/tabs/tab/__internal__/tab.context.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+export interface TabContextProps {
+  setError?: (childId: string, error?: boolean | string) => void;
+  setWarning?: (childId: string, warning?: boolean | string) => void;
+  setInfo?: (childId: string, info?: boolean | string) => void;
+}
+
+export default React.createContext<TabContextProps>({});
